LPS-86902 Make getDropSectionPosition a utils function
